chore: make tests running for dev build and update marks
