+++ test/e2e/configs/system.py
@@ -7,6 +7,5 @@ def get_platform():
 
 
 DISPLAY = os.getenv('DISPLAY', ':0')
-
 TEST_MODE = os.getenv('STATUS_RUNTIME_TEST_MODE')
