+++ test/e2e/conftest.py
@@ -1,5 +1,4 @@
 import logging
-from datetime import datetime
 
 import os
 import allure
