+++ test/e2e/tests/communities/test_communities_categories.py
@@ -22,6 +22,7 @@ pytestmark = marks
     'second_channel_description, second_channel_emoji',
     [pytest.param('Category in general', True, 'Channel', 'Description', 'sunglasses', 'Second-channel',
                   'Description', 'sunglasses')])
+@pytest.mark.critical
 def test_create_edit_remove_community_category(main_screen: MainWindow, category_name, general_checkbox, channel_name,
                                  channel_description, channel_emoji, second_channel_name, second_channel_description,
                                  second_channel_emoji):
