+++ test/e2e/tests/communities/test_communities_mint_owner_token.py
@@ -48,7 +48,7 @@ def test_mint_owner_token(keys_screen, main_window, user_account):
         if configs.system.get_platform() == "Darwin":
             next_view.start_using_status()
         SplashScreen().wait_until_appears().wait_until_hidden()
-        if not configs.system.TEST_MODE:
+        if not configs.system.TEST_MODE and not configs._local.DEV_BUILD:
             BetaConsentPopup().confirm()
 
     with step('Enable creation of community option'):
