+++ test/e2e/tests/onboarding/test_onboarding_negative_scenarios.py
@@ -52,7 +52,7 @@ def test_login_with_wrong_password(aut: AUT, keys_screen, main_window, error: st
         if configs.system.get_platform() == "Darwin":
             next_view.start_using_status()
         SplashScreen().wait_until_appears().wait_until_hidden()
-        if not configs.system.TEST_MODE:
+        if not configs.system.TEST_MODE and not configs._local.DEV_BUILD:
             BetaConsentPopup().confirm()
 
     with step('Verify that the user logged in correctly'):
