+++ test/e2e/tests/onboarding/test_onboarding_syncing.py
@@ -89,7 +89,7 @@ def test_sync_device_during_onboarding(multiple_instances):
             if configs.system.get_platform() == "Darwin":
                 AllowNotificationsView().start_using_status()
             SplashScreen().wait_until_appears().wait_until_hidden()
-            if not configs.system.TEST_MODE:
+            if not configs.system.TEST_MODE and not configs._local.DEV_BUILD:
                 BetaConsentPopup().confirm()
 
         with step('Verify user details are the same with user in first instance'):
