+++ test/e2e/tests/settings/test_settings_back_up_seed.py
@@ -20,7 +20,7 @@ def test_back_up_seed_phrase(main_screen: MainWindow):
         settings = main_screen.left_panel.open_settings()
         assert driver.waitFor(lambda: settings.left_panel.settings_section_back_up_seed_option.wait_until_appears,
                               configs.timeouts.UI_LOAD_TIMEOUT_MSEC), f"Back up seed option is not present"
-        if not configs.system.TEST_MODE:
+        if not configs.system.TEST_MODE and not configs._local.DEV_BUILD:
             assert BackUpSeedPhraseBanner().does_back_up_seed_banner_exist(), "Back up seed banner is not present"
             assert BackUpSeedPhraseBanner().is_back_up_now_button_present(), 'Back up now button is not present'
 
@@ -30,6 +30,6 @@ def test_back_up_seed_phrase(main_screen: MainWindow):
 
     with step('Verify back up seed phrase banner disappeared'):
         assert not settings.left_panel.settings_section_back_up_seed_option.exists, f"Back up seed option is present"
-        if not configs.system.TEST_MODE:
+        if not configs.system.TEST_MODE and not configs._local.DEV_BUILD:
             BackUpSeedPhraseBanner().wait_to_hide_the_banner()
             assert not BackUpSeedPhraseBanner().does_back_up_seed_banner_exist(), "Back up seed banner is present"
