+++ test/e2e/tests/wallet_main_screen/wallet - footer actions/test_footer_actions_send_nft.py
@@ -51,7 +51,7 @@ def test_wallet_send_nft(keys_screen, main_window, user_account, tab, receiver_a
         if configs.system.get_platform() == "Darwin":
             next_view.start_using_status()
         SplashScreen().wait_until_appears().wait_until_hidden()
-        if not configs.system.TEST_MODE:
+        if not configs.system.TEST_MODE and not configs._local.DEV_BUILD:
             BetaConsentPopup().confirm()
 
     with step('Set testnet mode'):
