+++ test/e2e/tests/wallet_main_screen/wallet - right click out of account area/test_right_click_manage_watched_address.py
@@ -15,7 +15,6 @@ from gui.main_window import MainWindow
 pytestmark = marks
 
 
-@pytest.mark.critical
 @allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703100',
                  'Manage a watch-only account from context menu option')
 @pytest.mark.case(703100)
